Move allowance related components to dedicated directory.
